move helper function and update logging level
